# examples/gameroom/daemon/openapi.yml
...
openapi: '3.0.0'
info:
  version: 0.3.17
  title: Supplychaind API
  description: REST API for Supplychaind
servers:
  - url: http://localhost:8000
paths:
...
  /xo/{circuit_id}/games:
    get:
      tags:
        - XO Games
      description: List all XO games within the specified circuit
      parameters:
        - name: circuit_id
          in: path
          description: Circuit id of the XO games to list
          required: true
          schema:
            type: string
      responses:
        200:
          description: list of XO games
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApiXoGame'
                  paging:
                    $ref: '#/components/schemas/Paging'
        500:
          description: Internal server error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        503:
          description: Server unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Error:
      additionalProperties: false
      properties:
        code:
          description: The HTTP error code
          type: string
          example: 404
        message:
          description: A message describing the error that occurred
          type: string
          example: Resource not found
      required:
        - message
...
    Paging:
      type: object
      properties:
        current:
          type: string
        offset:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        first:
          type: string
        prev:
          type: string
        next:
          type: string
        last:
          type: string
      example:
        current: /nodes?offset=10&limit=10
        offset: 10
        limit: 10
        total: 50
        first: /nodes?offset=0&limit=10
        prev: /nodes?offset=0&limit=10
        next: /nodes?offset=20&limit=10
        last: /nodes?offset=40&limit=10
...
    ApiXoGame:
      type: object
      properties:
        circuit_id:
          type: string
        game_name:
          type: string
        player_1:
          type: string
        player_2:
          type: string
        game_status:
          type: string
        game_board:
          description: string representation of game board, using comma-separated values format
          type: string
        created_time:
          description: time in seconds since the unix epoch the XO game was created
          type: integer
        updated_time:
          description: time in seconds since the unix epoch the XO game was updated
          type: integer
...